Update main-code.py
In this modified version, the souls_positions attribute is added to the PacmanGame class and initialized as an empty list in the constructor. It is then updated in the souls_move_randomly method.
By making this modification, the AttributeError should be resolved, and the code should be able to access the souls_positions attribute correctly.