Update main-code.py
In this modified version, the game_board attribute is added to the PacmanGame class and initialized as an empty list in the constructor. It is then updated in the initialize_game_board method.
By making this modification, the AttributeError should be resolved, and the code should be able to access the game_board attribute correctly when displaying the game board.